+++ contracts/blockGasLimit-dgp.sol
@@ -1,10 +1,10 @@
 pragma solidity 0.5.8;
 
 contract blockGasLimit {
-    uint256[1] _blockGasLimit = [
+    uint64[1] _blockGasLimit = [
         40000000 //default block gas limit
     ];
-    function getBlockGasLimit() public view returns (uint32[1] memory) {
+    function getBlockGasLimit() public view returns (uint64[1] memory) {
         return _blockGasLimit;
     }
 }
