+++ contracts/transactionFeeRates-dgp.sol
@@ -1,7 +1,7 @@
 pragma solidity 0.5.8;
 
 contract transactionFeeRates {
-    uint256[3] _transactionFeeRates = [
+    uint64[3] _transactionFeeRates = [
         10E8, //0: minRelayTxFee
         10E8, //1: incrementalRelayFee
         30E8  //2: dustRelayFee
